Extract more source files for better code navigation
