+++ app/src/main/cpp/CMakeLists.txt
@@ -24,7 +24,9 @@ add_library( # Sets the name of the library.
              SHARED
 
              # Provides a relative path to your source file(s).
-             native-lib.cpp )
+             native-lib.cpp
+             CpuThreadControl.cpp
+             DotProd.cpp )
 
 # Specifies libraries CMake should link to your target library. You
 # can link multiple libraries, such as libraries you define in this
