+++ app/src/main/cpp/CpuThreadControl.h
@@ -1,0 +1,5 @@
+#pragma once
+
+void SetCurrentThreadAffinity(unsigned int affinity);
+unsigned int GetLittleCoreAffinity();
+unsigned int GetBigCoreAffinity();
