+++ app/src/main/cpp/DotProd.h
@@ -1,0 +1,9 @@
+#pragma once
+
+int dotProductScalar(short* inputArray1, short* inputArray2, short len);
+int dotProductNeon(short* inputArray1, short* inputArray2, short len);
+int dotProductNeon2(short* inputArray1, short* inputArray2, short len);
+int dotProductNeon3(short* inputArray1, short* inputArray2, short len);
+int dotProductNeon4(short* inputArray1, short* inputArray2, short len);
+int dotProductNeon5(short* inputArray1, short* inputArray2, short len);
+int dotProductNeon6(short* inputArray1, short* inputArray2, short len);
